net/mlx5: explicitly include autoconf file on lib glue
The glue libarary uses the MACRO defined on mlx5_autoconf.h, hence
the file should be explicitly included on mlx5_glue.h, otherwise
the header file we re-define existing structs or enums.

It is not happening currently because many of the c files which include
mlx5_glue.h file also includes the mlx5_autoconf.h file.

Cc: stable@dpdk.org

Signed-off-by: Shahaf Shuler <shahafs@mellanox.com>