net/mlx5_vdpa: report virtio notify area
Select the notify area for the given virtqueue doorbells. The offset
selected is relative to the base address of the virtio doorbells on the
device PCI bar.

The current KVM and QEMU code only supports mapping of areas in size of
page, as it much fits into the way KVM register the EPT/Shadow page
table entries.

This is OK for BlueField1 solution, however, as an action item will need to
check if we can modify KVM and QEMU and support smaller sizes.

Signed-off-by: Shahaf Shuler <shahafs@mellanox.com>